Split shell shared key isolation tests
